Backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.8
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-elasticsearch-dsl==8.0
django-filter==25.1
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
elastic-transport==8.17.1
elasticsearch==8.18.0
elasticsearch-dsl==8.17.1
idna==3.10
kombu==5.5.2
msgpack==1.1.0
mysqlclient==2.2.7
pillow==11.2.1
prompt_toolkit==3.0.50
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
qrcode==8.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
